pass needed service info to service application component
